+++ src/main/java/domain/deck/Type.java
@@ -1,6 +1,7 @@
 package domain.deck;
 
 public enum Type {
+
     ACE(11, "A"),
     TWO(2, "2"),
     THREE(3, "3"),
