+++ src/main/java/domain/result/WinningResult.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 public enum WinningResult {
