+++ src/test/java/domain/result/WinningResultTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.result;
 
 import static org.assertj.core.api.Assertions.*;
