+++ src/test/java/domain/user/DealerTest.java
@@ -11,7 +11,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
 
-import domain.WinningResult;
+import domain.result.WinningResult;
 import domain.deck.Card;
 import domain.deck.Symbol;
 import domain.deck.Type;
