+++ src/test/java/domain/user/PlayersTest.java
@@ -12,7 +12,7 @@ import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
 
-import domain.WinningResult;
+import domain.result.WinningResult;
 import domain.deck.Card;
 import domain.deck.DeckFactory;
 import domain.deck.Symbol;
